glovera-backend/requirements.txt
fastapi 
pydantic 
uvicorn 
python-multipart 
python-jose 
requests
pymongo
pycryptodome
python-dotenv
firebase-admin
pydantic[email]
groq